MB-61387 Track markFirstSnap per keyspace in stream reader
The markFirstSnap variable is used to identify if for
any stream, first snap optimisation has to be enabled
when streams are starting to stream with seqno. 0.

Currently, this variable is tracked per stream. If the
first keyspace to open the stream enables first snap
optimisation, then it will be applied to all keyspaces
that follow in the same stream even if it is not supposed
to be enabled. This will lead to plasma/memDB skipping
back-index look-up for mutations happening in the first
snapshot.

If an index has non-zero items and it processes a different
first snapshot (due to stream restart) than the one with which
it was build and the stream for that index restarts from "0",
then duplicate rows will be inserted into the index. This
will lead to incorrect results.

The fix for the issue is to track the "markFirstSnap" value
for each keyspace instead of tracking for entire stream.

Change-Id: I3ac37b31b5ae35613630c2663f2a7cf05cb62d2e