MB-61140 Add support for CreateIndex6 API
A new API CreateIndex6 is introduced by GSI client.
This API accepts 3 new arguments:

a. hasVectorAttr - A slice of boolean values. If any index
   field has VECTOR attribute, corresponding entry in
   hasVectorAttr becomes true

   A slice of boolen values is considered instead of a
   single integer representing the vector position for
   2 reasons:

   a) VECTOR is an attribute like DESC. Hence, it is being
      tracked similar to DESC
   b) When there is a flattened array before VECTOR expression,
      the relative position of the vector changes from the original
      index definition as flattened keys will be exploded. This
      can lead to unnecessary confusion. Other components
      (e.g. projector) will use this array will deduce the position
      of the vector field inside the index.

b. include - List of secondary expressions that needs to be put
   in include columns for BHIVE index

c. isBhive - A boolen value indicating the support for BHIVE
   indexes

A new version INDEXER_VEC_VERSION is added. Since the release
vehicle is unknown at this point, INDEXER_VEC_VERSION is used as
a placeholder.

Change-Id: I9321512588a1dc7cca09638fa778b2c809d8c7c9