MB-61140 Extract dimension, similarity from with clause
"dimension" and "similarity" are some of the mandatory
parameters when creating a vector index.

"dimension" represents the size of each vector that will
be present in the incoming document. If a vector in a
document does not have same length as "dimension" value,
then a NULL value will be inserted

"similariy" represents how 2 vectors are computed. It should
be one of "EUCLIDEAN", "DOT_PRODUCT", "COSINE_SIM", "L2"
(Case in-sensitive)

Change-Id: Idb2be30dfa418a1f25150e38c661bc2aaf7e103c