MB-61140 Update vector metadata fields in index definition
Include expressions and vector metadata are added to index definition.

Include expressions are tracked separately from vector metadata as
it is possible to have explicit include expressions for a normal
index as well. Although such a feature does not exist today, it
can be added in future. Hence, it is de-coupled from index metadata

Change-Id: Ieb004db5969e6d08500493e48e6b13e8415e8ccb