[BP 7.6.2] MB-61824: Reduce redundant logging for decodeTransferToken
When a new ShardTransferToken is observed, print the IndexInsts of
the TT only when the Node is either the Master, Source or Destination.
If the node is not relevant for processing the TT, even more compact string with only the TT's basic information is printed.

NOTE: When the token is created, a more verbose information about the
token is printed on every Node, to ensure that atleast once the entire
info of the TT is printed on each node.

Link for comparison: https://drive.google.com/drive/folders/1WgkZPRr9EDlPLZZj9kEf4OSgXsXN6C2S?usp=sharing

Change-Id: I8b551295919270dbef453b28afcff14980a728a8