MB-61139 Add vector related information in PROTO request to projector
The `hasVectorAttr`, `dimension` and `include` fields are sent
from KV sender to projector as a part of PROTO request.

IndexEvaluator in projector uses this information to deduce the
following fields:

a. isVectorIndex - Set to true of any of the secondary expressions
   are defined with VECTOR attribute

b. vectorPos - Is set relative to the position of VECTOR attribute
   in skExprs. If the VECTOR attribute is inside a flattend array
   expression, then vectorPos will be set to the position of
   flattened array expression. Another variable
   `vectorPosInFlattenedArray` is used to track the position of
   the vector attributed inside flattened array expression

c. vectorPosInFlattenedArray - Set to the relative position of the
   VECTOR attribute in flattened array

d. dimension - Expected value of the vector dimension from incoming
   documents

e. include - List of secondary expressions that needs to be evaluated
   for include columns

Change-Id: Ibb4a1dfa2536d7fb291fb9d49d6ea3ee58b50e8d