MB-61139 Add validateVector method
validateVectorMethod does 3 types of validations:

a. Incoming value has to be a vector
b. The size of the vector has to be equal to the
   dimension of the vector specified in index definition
c. All entries of the vector are float64 values

If a vector is valid, then it is converted to float32
values and passed onto indexer

Change-Id: I67f3ef1ccbe44d8d67c35825ec926862d1bbe5f9