MB-61771 Add support for with "nodes" clause and shard affinity
BP of MB-60735 to 7.6.2

This patch adds the support for with "nodes" clause when
shard affinity is enabled.

To support with "nodes" clause, planner will try to find
a common slot across all the target nodes for which the
partition is going to.

When with "nodes" clause is used, planner will not retrieve
the information from entire cluster. The information will
be retrieved only from the nodes specified with "nodes" clause.
Because of this, the slot distribution can be incomplete.

E.g., a slot can exist on nodes n1, n2 while planner wants
to place on nodes n0, n1. Planner does not know about the
presence of the slot on n2. If it uses the slot for placement
on n0, then it will lead to incorrect placement as the slot
with same replicaId can exist on 2 different nodes.

To avoid such cases, planner will try to find a slot that is
common to all target nodes. If none is found, a new slot
is created and shared with all the target nodes. Further
index creations on these set of nodes will try to reuse the
slot.

Change-Id: I8213fd21f639ae8de6bbbdf582ae305350e46dbb