MB-61779 Move indexer version to 8.0 in cluster_info
As master is running on 8.0, move indexer version to 8.0 so
that features like file rebalance can work seamlessly.

Since there is a possibility of 7.7 version at the time of
this patch, linear numbering scheme will not work for 8.0
version i.e. we cannot move indexer version from 9 to 10 as
7.7 should come before 8.0. Hence, using a large number like
800 can solve the problem so that 8.0 can be 800, 8.1 can be
801, 8.2 can be 802. This scheme assumes that the minor
versions do not exceed 8.99 (or) 899. When 7.7 releases,
we can assign indexer version as 707.

Change-Id: I6c415c2372eabe9512291cd6a2c1e18db8cf4559