MB-61139 Add N1QLTransformForVectorIndex method
For vector index, few additional checks are to be made
when parsing the incoming documents. To separate the
logic between parsing a document for a scalar index
and a vector index, a new method is added.

This patch is just a code re-ordering. Actual logic
will come in later patches.

Change-Id: If474c6af44ecf4f1c97a273b76b908be2b4b2942