MB-61139 Integrate N1QLTransformForVectorIndex into mutation processing path
Projector evaluates the incoming document for expressions
with VECTOR attributes. Projector will follow the below rules:

a. If VECTOR entry is MISSING/NULL and it is leading key, the
   document will not be indexed

b. If VECTOR entry is MISSING/NULL and it is non-leading key,
   we index MISSING/NULL for the VECTOR entry

c. If the VECTOR entry is an invalid and it is leading key,
   the document will not be indexed

d. If the VECTOR entry is invalid and it is non-leading key,
   the VECTOR entry will be indexed as NULL

Change-Id: I3596dc964904d0fb3b4d688a1f9ed3f3b8b859fa