MB-61139 Add vector data to key versions
For each vector index, an array of vectors (i.e. array of
array of floats) is added in KeyVersions struct and it will
be passed onto indexer.

Indexer will use these array of vectors to compute centroidIDs
and replaces the dummy centroids in incoming mutations with the
actual centroidId values

Change-Id: Icd57e09595c085723dee4595063705bd3cc70818