MB-61090 Implement serialize/deserialize methods for Codebook
Serialize method marshals the codebook to a []byte. The caller can take
the []byte and use it to persist the codebook to the disk.

Deserialize method takes []byte as input, unmarshals the data and
initializes the codebook. This method can be used for recovering the
codebook from its persisted state.

Change-Id: I211c70d26199058b4997e96f3d95b761829e13f9