MB-63381 Remove tombstone instances from topology
BP of MB-61861 to 7.6.4

The current logic to remove tombstone instances from topology
has a bug in which it compares the instId of the tombstone
and the realInstId of the  partition. Instead, the realInstId
of the tombstone instance and the realInstId of the partition
should be compared.

Also, once all the partitions of an instance are removed, the
tombstone instance is not removed from topology. This can cause
an instance leak in the topology. This patch fixes both the
issues.

Change-Id: I448268fba78adce5cdf00afbf9f695977ae02166
(cherry picked from commit b04e652fb335a4f3ec38eacf3b52df00283f59f5)