MB-61861 Remove tombstone instances from topology
The current logic to remove tombstone instances from topology
has a bug in which it compares the instId of the tombstone
and the realInstId of the  partition. Instead, the realInstId
of the tombstone instance and the realInstId of the partition
should be compared.

Also, once all the partitions of an instance are removed, the
tombstone instance is not removed from topology. This can cause
an instance leak in the topology. This patch fixes both the
issues.

Change-Id: I448268fba78adce5cdf00afbf9f695977ae02166