MB-61911 Fix active connections book-keeping in queryport client
BP of MB-61872 to 7.6.4

Currently, the active connections counter is incremented after
a new connection is genereted. When multiple scans are fetching
a connection, then all of them write to createsem channel first,
create a new connection and then increment the counter.

By the time the counter is incremented for all connections, few
connections could already be returned to the pool - This leads
to a case where the count of active connections becomes inaccurate
(or) does not reflect the actual length of createsem channel.

This patch address the issue by incrementing the counter before
fetching the connection so that by the time a connection is
returned, the counter already accommodates for the connection
and therefore decrementing the counter will be coherant with
the length of createsem channel.

Another issue is observed with releaseConns go-routine.
In this go-routine, when connections are released, the
createsem channel is not drained. This can lead to a case
where the createsem channel is almost always full and only
"minPoolSizeWM" connections are maintained eventually. It
is incorrect behaviour as the pool can grow till 5030
connections but we are limiting only till 1030 connections.

To address the issue, the createsem channel is drained
when the connection is released. The draining is perfomed
in a non-blocking way so that the go-routine is not blocked
even if createsem is empty (which ideally should not happen).

Change-Id: Ibbbc582343f2574ca8dda93d5510c0ae0bb2aa1a
(cherry picked from commit 3ba3eb8c81e80ca427a82ecfcd4022c1e1ba1cb9)