MB-61914 [BP 7.2.6]: Limit number of threads in persistence for MOI
The config - indexer.settings.moi.persistence_threads - is
expected to limit the CPU consumed by disk snapshotting. By
default it is set to half of GOMAXPROCS(0). This config limits
the CPU consumption in 2 ways by controlling:

1) The number of threads allotted to each index for disk
snapshotting: This is proportional to the fraction of total
node items in that index. Each index should get at least 1
thread.
2) The number of indexes for which disk snapshotting can
happen in parallel - capacity of moiWriterSemaphoreCh.

So, if there is a mix of large and small indexes, due to the
minimum 1 thread, there may be more than intended threads
running disk snapshotting.

To handle this, use moiWriterSemaphoreCh as a gate for the
number of threads running instead of the number of indexes.

Change-Id: I47eb5574d2f67497b7c11f1347dd6b855ac15c6a