MB-61139 Add codebook to plasma_slice
Each plasma slice will hold a codebook. It gets initialized
at the time of index build. It is possible that the "nlist"
parameter (i.e. the number of centroids required for IVF
indexes) is not specified at index creation time. In such
cases, indexer has to deduce the "nlist" value based on the
items count in the keyspace.

Since nlist has to be known at the time of codebook initialisation,
InitCodebook() is expected to be called only during build phase
of the index. If any error is observed during the training phase,
then indexer is expected to RESET the codebook so that next build
attempt will create a new one.

Change-Id: I5549f9bc49840803be3261aec742b330721f9f82