MB-61126: Perform codebook transfer as part of processShardTransferMessage
Perform all codebook transfers associated with the Shards, the
processShardTransferMessage function now performs the codebook transfer
first and after all successful codebook transfer it proceeds with shard
transfer.
If any codebook transfer fails. The rebalance is considered a failure.

The codebook transfers reuses the StorageMgr Cancel and RespCh, it only
closes the StorageMgrRespCh when a rollback (cancel ch closed) during
the Codebook transfers. Else it doesn't close the RespCh. The RespCh is
still the responsibility of processShardTransferMessage() after
successful ShardTransfer

Change-Id: I904cd119693c096f885cb1cc57307705cb5a8b38