MB-61139 Compute centroids for instances during index build
If an index is defined without the centroids for IVF, then
the number of centroids are computed at index build time
based in the items count in the keyspace. As each index
instance can be built at different points in time, the
centroids required for vector indexes are computed at
build time so that we can get an accurate number based on
the items count in the index at the time of build.

Centroids are derived from items count based on the
following logic:

a. If the itemsCount of a keyspace is < 1000000 (1M), then
   numCentroids = itemsCount / 1000
b. If the itemsCount of a keyspace is > 1000000 (1M), then
   numCentroids = sqrt(itemsCount)

If itemsCount is < 1000, then only one centroid will be used
for training the index. Index build would fail if there is
any error observed while retrieving the items count from
data service.

Change-Id: If5f3156f91cb0686c4f669c0b1b4a24a954dd259