MB-61139 Start training for vector indexes
For vector indexes, indexer will start training asynchronously
in a go-routine.

On a successful training completion, the index state is moved to
TRAINING_COMPLETED. If training fails for any reason, then
the index state is reset from TRAINING_IN_PROGRESS to default
value and the state of the instance is moved to error.

Change-Id: I33ca85b17f919c68df03e904a6a4922971f6d369