MB-61139 Handle runtime errors during vector index builds
A vector index build is different from normal index build
as there can be runtime errors due to insufficient items
for training the index in the bucket. This can happen
either due to the bucket being empty or the bucket does not
have enough documents to train the index (i.e. number of
documents are less than the centroids defined for vector
index).

In such cases, user intervention is required. For non-deferred
indexes, the index/instance will be deleted. For deferred
indexes, indexer will still keep the metadata information
and update the error to the instance in metadata. Indexer
will skip retrying the index build for training errors.

Change-Id: I256291342c87be700edb6ddabe7ec4bd16109690