MB-61139 Use TrainingPhase instead of isTrained
For each index instance, a TrainingPhase variable is added.
This variable replaces the "isTrained" boolean value as
index training can exist in three phases:

a. TRAININIG_NOT_STARTED: Training is not started for the
   index

b. TRAINING_IN_PROGRESS - Training has started for the batch
   of indexes

c. TRAINING_COMPLETED - Training has completed for the index
   instance

The TrainingPhase value moves to TRAINING_IN_PROGRESS at the
start of training. Once training is successful, the state will
change back to TRAINING_COMPLETED.

UI will use this information to populate the status on UI.
For vector indexes, "training" will be displayed and for
non-vector indexes "waiting for training completion" will
be displayed on UI

Change-Id: If22c6b0475793771affac47ce18f87d866c812dd