MB-61095 Add RandomScanner for KV doc sampling
RandomScanner is a new interface and its corresponding implementation
which provides a mechanism for random sampling of KV docs. Given an
input collection and sample size of required documents, it utilizes
memcached random scan protocol to sample the KV documents.

This works with both couchstore and magma bucket types and allows
non memory resident documents to be sampled as well.

Change-Id: I68e4edad2087fc1b5bcf87799dc41839e96e2105