MB-61139 Add support for fixed length encoding scheme
When projector sends a mutation for vector index, it will
encode a dummy centroidId with value of -1. This has a
drawback: If indexer has to replace the centroidId, then new
memory has to be allocated for the key as the length of the
encoded centroidId for the vector and the length of dummy
centroidId can be different.

With fixed length encoding, all centroidIds will have constant
length (8 bytes) as the centroidId is encoded as hexadecimal
string. This means that if indexer know about the position of
vector field inside the collateJSON encoded key, it can replace
the new centroidId in-place without generating new memory for
the incoming key

Change-Id: I25f30d572b458cafa1945beb46a432360f20f0cd