MB-61139 Skip indexing invalid vectors fields
When vectors are not valid - either due to dimensionality
mismatch or due to vector not being present, then skip indexing
the document irrespective of whether the expression with VECTOR
attribute is leading key or not.

A document will be indexed on an index with VECTOR attribute only
when the vector is valid and it exists. Documents with MISSING (or)
invalid vector values will be skipped from indexing.

Change-Id: Ib664bef005daf251e44327143fc5abd9e8d075de