MB-61095 Add utility method to fetch random KV sample
Add a utility method in common to fetch random KV sample.

FetchRandomKVSample fetches a given sampleSize number of documents
from a collection. It returns "datach" on which sampled documents are
returned in DcpEvent format. Any error during sampling is sent on
the err channel. datach is closed by the sender when all the sample
docs have been sent.

Change-Id: Ied7c58ca5c28cad66a8db9f175e6414392082b6f