MB-61126: Introduce copy config struct to pass config to N2NCopier
Introduce a CopyConfigMeta to keep information required to instantiate
a node to node copier for codebook.
This is required as makeFileCopier will be called at the time of
codebook transfer as well as codebook cleanup. And hence a generic
struct needs to be passed as an argument to the constructor method

Change-Id: Icf562602a0d8cede4cf497b28a2ae2e85ea8e632