MB-61126: Perform codebook cleanup as part of processTransferCleanupMessage
For any errored out shard transfer, cleanup of the codebooks associated
with the TT are also performed as part of processing the
MsgShardTransferCleanup.
For the cleanup, a file copier is initialised and then cleanup for each
of codebook in the staging folder is performed. This step ensures the
codebook folders are entirely removed from the staging folder

Change-Id: I82d2825933fb48c72556e32f71a7b9ed12fa437b