MB-100 make grep read binary files too in dotest
bg: due to the encoding scheme used on the new CI
it is happening that sometimes the grep command
believes that the test.log is a binary file and does
not log the 'FAIL'ed test it matches from the file.
this is leading to CI giving +2 even if some tests
are failing.

fix: make grep outputh the matches even if it feels the
input is a binary file.

future work: currently from my experiments, it is
observed that the grep exits with zero code even
when it sees a binary file has matched. so it is
not possible to fail CI if grep fails. this is kept
as an open item

Change-Id: I1ee8ae4b5401855d0ba17eeae536198b57390252
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>