MB-61030 [BP 7.6.2] Evaluate align flag again after TS update
ensureMonotonicTs can update a timestamp seqno/snap information if it
finds any constraint violation. It should evaluate the snap flag again
in such a case as updating the seqno/snap information can lead to a
change in timestamp being snap aligned or not.

In case of this MB, due to dynamic changing of smallSnapshotThreshold,
it is possible that some of the later TS has lower seqno than the
previous one e.g.

1.smallSnapshotThreshold=10, s[0]=0,s[1]=20,seq=15
TS1=[0-20,15] - doesn't qualify for smallSnapThreshold

2.smallSnapThreshold changes to 20
TS2=[0-0,0] - qualifies for smallSnapThreshold again
TS2 is smaller than TS1

ensureMonotonicTs will update the TS2 with the seqno/snap information of
TS1 but will not set the snap aligned flag correctly.

Change-Id: I2814313c23305cd6e8c29ecc8e5596451c88b87d