MB-61139 Persist codebook to disk on successful training
After training is completed, indexer will try to persist the
codebook to disk. Once the persistance is completed, the
state of the instance is updated to TRAINING_COMPLETED in
metadata. After the state is updated, then indexer would
start the build.

If any errors are observed during the training phase or
during the persistance phase, then the index build will
not proceed further. Error value is updated in the instance
metadata.

Change-Id: Ida034d1cb001bea8694c27eff20ee85b40649bed