MB-61139 Recover codebook from disk on indexer restart
After indexer restarts, plasma slice tries to recover the
codebook from disk. During recovery, it can encounter two
types of errors:

a. Codebook not found - This can happen when the index is
created but not build (or) indexer restarted before persisting
codebook to disk

b. Codebook corrputed - If the codebook is partially persisted
or the storage file on disk is corrupted, then this error is seen.

Indexer tries to handle these errors in the following way:

a. If the instance is not trained, then codebook not found error
is returned from the slice initialization. Indexer would ignore
the error as not finding codebook is expected

b. If the instance training is in progress while indexer restarted,
and codebook path is not found, then indexer must have restarted
before persisting the codebook. Indexer would ignore the error.

Index build retry (due to the presence of build command token) will
attempt to retrain the index.

c. If the codebook is found and it is valid while the training phase
of the instance is TRAINING_IN_PROGRESS, then indexer would successfully
recover the codebook for the slice. During build retry (by build command
token), the metadata state of the instance is updated to TRAINING_COMPLETE

d. If codebook is corrupted during recovery, then the index will be removed
and the metadata is cleaned up. Index will be considered as failed partition.
Rebalance (or) drop and recreate of index is required to fix this state

Change-Id: I1d3fd02592d8f084617edad7e6e6258b2f2e373d