MB-61139 Skip using 4 bytes for len(sha_encodings) for non-array composite vectors
For non-array composite vector indexes, indexer expects (and processes)
only one vector per document. Therefore, there is no need to include
4 additional bytes in storage to capture the number of SHA encodings
or the number of quantized codes in back index and main index respectively.

Indexer will decide if the VECTOR attribute is on array expression or not
and act accordingly.

Change-Id: I42278702859ed23c2592333d59f838fc2ec6a72f