MB-61126: Add functionality for RestoreCodebook and util functions
Once the codebook has been transferred to the temporary staging
directory(ShardRestore is complete) at the destination node, the
codebooks need to be moved to their respective IndexPath folder on the
destination node.
As part of the Codebook restores, the files need to be only copied to
the relevant folder.

This patch focuses on introducing the function that will perform the
actual move from the srcPath to the specified destPath, for the codebook
of the IndexInst passed

There is also an introduction of checking the respMsg type in the
startShardRestore, which will be relevant for processing responses sent
during the actual codebook restores - relevant in the next patch

Change-Id: If914cf2575c4afc756c8b9fbd49476d50aa34c15