MB-61126: Perform codebook restore as part of processShardRestoreMessage
Once all the shard are restored by plasma, the relevant data for the
shards has been restored in the relevant production directory on the
destination node.
Sequentially after perform the codebook restore for the indexes
associated with these Shards.

Now the startShardRestore waits for a response from the Codebook Restore
before exiting and changing the state of the TT to
ShardTokenRecoverShard.
For Shards with no vector indexes, a response from codebook restore is
also returned but the errMap and CodebookPaths are empty.

Change-Id: I77b12208c352f0501c85232917400acb75cdd540