MB-61149: Fix parsing on quoted string in cbindex batch process - indexDDL
When -index_ddl is called directly, the os.Args trim the quotation marks input as well as keep the quotes string as a single input.
But when batch process happens, after reading from the file, the
entire line is split into substring based on whitespace character, which
results in breaking up the CREATE INDEX Statement into multiple input
substring.

Now the quoted texted is considered together as a single input substring
before passing to parseArgs. And the quotation marks are trimmed from
the indexDDL value

Change-Id: Iead196d84fbd788d5d47b044285ba558c351c9a4