MB-62761 register service manager with ns_server before bootstrap
ns_server needs indexer to register the service manager
before we use cluster information during bootstrap.

this patch makes the service registration happen before
bootstrap and then continues to bootstrap. all the service
managers are initialised async. The RPC calls have a wait
for the init to finish with a timeout. if we hit the
timeout then we will return the following error to ns_server -
    -> "indexer still in bootstrap"

Change-Id: Idca1df95526e0564fe13fffa45228a662e42c256
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>