MB-62929 Process all createCommandTokens before proceeding with drop
BP of MB-62865 to 7.6.4

During alter index, the createCommandToken posted in metaKV will
have the defnId-requestId format where requestId is a non-zero
uint64 value.

According to the current protocol, when there are both create and
drop instance tokens in metaKV, it is expected that the create
command token is removed first and then drop command token is
removed. To facilitate this, the cleanupDropInstanceCommand()
method checks for the presence of a create command token in
metaKV.

However, this check only checks for tokens with defnId as key
and skips all tokens with defnId-requestId as keys. This means
that a drop instance token can be removed from metaKV while
a create command token exists for the same instance. This can
lead to a case where the dropped instance is re-created and
re-built again which can distrupt the resources of a running
system.

This patch fixes the issue by reading all create command tokens
for a particular defnId and then process drop only if a create
command token with the instanceId of drop command token does
not exist in metaKV. Otherwise, drop is skipped from processing.
in metaKV.

Change-Id: Iba902154569f73082054b7394daa1ba11cf46ea7
(cherry picked from commit 8e223baf94848764223c965ccf48bd10e3c6c53b)