MB-61139 Handle instance drop during training
This patch will handle any instance drop during training phase.
When training for an instance is in progress, indexer will
update a shard map with the details of index getting dropped.

The initiateTraining() method will check for dropped instances
from this shared map and skip those instances from further
processing. It will also send a  message to indexer main-loop
asking to cleanup the index data. As the index state for
instances in training will be CREATED/READY, no stream
management is required.

Change-Id: I3cbb417ce1a245c703f5417f176ef4b2f2a8f57a