MB-61139 Do not wait for training completion during drop
Consider 2 instances idx_1 and idx_2 being built in the same
batch. As they are built in same batch, training for these
instances will happen together.

While idx_1 is being trained, if idx_2 is dropped, then the
current code will wait for training of idx_1 to complete before
processing the drop. This path fixes this "implicit wait" by
processing drop immediately.

If idx_1 is undergoing training and it is dropped, then indexer
will wait for training completion before dropping the instance.

Change-Id: I1c5cec4ff88a95ddccee72b22038ba460d496246