MB-62935: Defer and issue build index req to avoid N1QL index DDL timing out
For Composite Vector indexes, the index creation can take a significant
time in the training and build phase. This can cause any N1QL queries
to time out and fail the test suite.
Added a function to execute deferred index creation and then waiting for
the index to become active

Due to the large amount of time taken for the Vector Indexes to become
active, the function that waits for the indexes is also updated to print
at an exponential backoff and then linear sleep if it reaches a
threshold(tenth of the indexActiveTimeout).

Change-Id: I30585306041dd6d6b7d24d8ff30e523f53eefa77