MB-63034 Add Iterate method for BHIVE slice
Iterate() method is the foundation on top of which all
other Snapshot reader methods are implemented. This
patch implements the Iterate() method on bhive snapshot
using KeyPrefixIterator.

The KeyPrefixIterator returns rawKey and rawMeta. rawKey would
be the documentID. rawMeta would include recordID and the
metadata given by index service to bhive storage. RecordID
will be used internally to fetch the actual Value stored for
the key. The support for the same will be added in future
patches.

Change-Id: I6d8a33b6bc75025d810f9de8a4fb565774c3245d