MB-63136 Check all droppedInsts before processing successMap
After training is completed for an instance, if it is dropped,
the instance would be a part of successMap but it will be
removed by "handleTrainingAndCheckForDrop" method. Once the
training is completed for all instances, this can lead to a
case where successMap will be processed for dropped instance.

Since the drop is already processed, the index will be removed
from book-keeping. So, when successMap is processed, indexer
does not know if the instance is dropped or not. This is leading
to a panic.

The fix for this issue is to check all the instances that were
dropped during training. This will ensure that indexer will
always have the necessary information to deduce if an instance
is dropped or not.

Change-Id: I58f0584e9bb02f242385d7320750d6b5e564efd3