MB-63127 Notify builder on training completion
After training completion, indexer will attempt to build
the indexes. However, if indexes could not be built for
any reason (like build already in progress on same
keyspace), current logic will not retry build as builder
book-keeping is cleared with the instances in training.

Hence, notify builder on training completion change so
that the indexes would be re-built. Also, fix a minor
typo

Change-Id: If1f6a44a8a048a4b29074b10534ae1cf766d84af