MB-63193 Consider all proxies of real instance when adding to MAINT_STREAM
BP to 7.6.2-MP2

Current logic for ADD_INDEX_LIST_TO_STREAM is built on an
assumption that all partitions of the index are built in
same batch. While this assumption is true for initial index
builds and DCP rebalance, shard rebalance does not honour
this assumption.

This leads to a case where partition index build can happen
in batches and the partitions added to the stream in first
batch will be over written by the partitions added to the
stream later. After rebalance, the partitions in first batch
will not receive any mutations from projector.

To address the issue, indexer will now consider all the proxy
instances that are in CATCHUP or ACTIVE state and add all those
instaces to MAINT_STREAM. This ensures that the real instance
is always updated with all the necessary partitions even if
the index is built in multiple batches.

Following tests have been performed to verify this patch:

a. Normal partitioned index build - This code path should
not be hit as realInstId is zero for normal partitioned
index build

a. DCP rebalance involving failover, swap cases where a
partitioned index moves between nodes. During DCP rebalance,
all proxies are built in same batch. Therefore, there should
not be any impact of this  patch on DCP rebalance.

b. Shard rebalance where each shard transfer token holds
a subset of partitions and inducing artificial sleep during
transfer so that each token takes arbitratry amount of time

Change-Id: I9cf2d8ec1db98ab238170a26dea9bfce3b764668