MB-63203: Have retryable Index build after training completion in Rebal
During Rebal there can be multiple builds going on, since the build for
vector indexes involve an asynchronous training path another build for
the same keyspace can be picked by Indexer.
Due to any other transient error (ongoing build etc) the build issued
for Vector indexes after the training can fail, furthermore these
the index build is not retried by builder as a rebalance is going on.
This causes the Rebalance tokens to not proceed and hang.

Introduced a retryable mechanism during handling of training done msg
this ensures during Rebal the build is retried for any transient errors

Change-Id: I866d3d09c1d2c7ae901d955156155aad78425806