MB-63279 introduce dummy shard transfer token
in an edge case, if we have no shard transfer tokens (usually
during upgrade) then we will not move the state of DCP transfer
token with IsPendingReady ahead; instead we finish the rebalance

this state transition now happens in cleanup by the destination
node. the problem here is that the source node may not see these
state transitions finishing the rebalance.

now a major issue here if the partition fails due to some reason
then the index becomes unavailable as the source node will be
gone from the cluster

this patch addresses the issue by ensuring we move the transfer
tokens ahead while the rebalance is going on rather than
this being a cleanup job
* master posts a dummy empty shard transfer token for the
 follower node which will have to do the state transitions
* the token will be in Merge state with no Insts
* when follower sees such a token, it checks if all the shard
 tokens are ready to merge. this check will pass and then
 follower starts partition merge for DCP tokens
* once follower is done, it will transfer the shard token
 state to ready from where master deletes the dummy token

this approach ensures mixed mode follower compatibility

Change-Id: I6c7bae67659256a6f1042a6bb0e9032687a704bc
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>