MB-63203: Update err in Inst's meta and error out transfer token
As async training is not returned back in the /buildIndexRebal REST
call's response, the err has to be observed after the response has
been returned.
After the handleBuildIndex response is returned back to the retryable
call, if the error is of non-transient nature, the idxInstMap is updated
with the error. If at the time of waiting for Index build if error the
non-transient error is found, Rebalance is failed and TT is errored out

The retry runs in a go-routine with the intance error update being
handover to the Indexer

Change-Id: I38d538d50ffec708ca47ebc82320ffe5fbedd70b