MB-61090 Allow index training without clustering
If number of centroids is equal to the number of input
vectors for training, then clustering is not required.
Add the centroids directly to the quantizer from the list
of input vectors. Faiss considers this as a valid case to
skip training the quantizer. The sub-quantizer training will
still happen.

Change-Id: I09ef6a40ca378b59e472d63e9fec2cad4416bc01