MB-63034 Add rerank config and parameters in scan request
For re-ranking, the number of rows scanned will be typically
"R" times greater than the actual limit in the scan.
E.g., if "limit 10" is issued in the scan, then R * 10 will be
the actual rows scanned. Re-ranking will be perfomed on these
"R * 10" rows. The variable "rlimit" captures this limit which
is introduced with this patch.

Change-Id: Ie8dae27a9a930030c7ee388bad0c7983fdda0ca7