MB-63136 Handle drop of instance that is under training
Currently, the drop of last insance in training batch is
not handled properly. This can lead to a panic during index
build after training is done.

This patch fixes the issue by skipping the build of last index
instance in the training batch incase drop is received while
training is in progress.

Change-Id: Id4b17aa8928e153f68eacc9a182da1056dc244d1