MB-63428 Set initial node from index whose initialAlternateShardIds are non-nil
BP of MB-63370 to 7.6.4

For a proxy index, there can be multiple grouped indexes with few
indexes requiring DCP rebalance and few indexes requiring shard
rebalance. For such a proxy index, currently the initialNode is
set based on the first non-nil initialNode in the list of
grouped indexs.

If the first non-nil initialNode index needs moving via DCP
rebalance, then the other indexes in the proxy will incorrectly
use the initialNode and assume that the shard is moving from
that initialNode (while it might be moving from a different node)

Instead, the initialNode of proxy index should be set based on
an index whose initialNode and initial alternate shardIds are
non-nil. This will ensure that shard proxies will have correct
initial node. initialAlternateShardIds being non-nil means that
the shard pair exists in the cluster prior to rebalance and
shard would be moving from such a node.

E.g., if n1:i1 and n2:i2 are moving to n3 and i1 needs DCP rebalance
while i2 needs shard rebalance, the initialNode of shard token
generated for n2:i2 -> n3:i2 movement should use n2 as initial node.
In the current code, n1 is being set as initial node for n2:i2 -> n3:i2
movement. The patch fixes the issue.

Another issue is with how shouldExcludeIndex method operates. For
replica repair cases, overrideExclude is set to true. So, the
initialNode is never checked for nil value in shouldExcludeIndex()
method inspite of it being nil.

With proxy index, a special case happens - The initialNode can be nil
in cases where all indexes of the proxy are moving via DCP rebalance but
overrideExclude is false as none of the indexes are undergoing repair.

To handle such cases, add a nil check for initialNode. Functionally, it
won't change anything for non-proxy indexes as overrideExclude = true
means that initialNode is nil and overrideExclude = false means
initialNode != nil. For proxy, overrideExclude = true if any of the
grouped indexes are undergoing repair. In such cases, the index is
not excluded from placement on the node. Otherwise, the proxy index
is excluded from placement on the node.

Change-Id: Ifd1f24d7f3f3416be4d1fbd48f8d4ee15682b89b
(cherry picked from commit db1dbdf9ca77013e101a26ac123b36ed52078f5e)